Fix UTF-16 decoding bug
It is known for certain that the player name is serialized in little-
endian, so the byte order must be explicitly specified when decoding
the bytes. Otherwise, this invokes the default behavior for UTF-16
decoding. In CPython, this goes as follows:

1. Check the first two bytes for the presence of the BOM (U+FEFF).
   If these first two bytes are ordered FEFF, then the stream is big-
   endian, so the UTF-16 decoder will decode the rest of the data as
   big-endian.
   Otherwise, if these bytes are ordered FFFE, then the stream is
   little-endian, so the rest of the data will be decoded as such.

2. If no BOM is found, the decoder defaults to using the system's
   native byte order to decode the data.

There are two obvious problems with this. First, under normal
circumstances, the player name will never contain a BOM at the start,
since there is no way to input such a character on the Switch. Thus,
under these conditions, the decoding behavior is platform-dependent.
Simply put, this will result in the data being decoded in big-endian
when the code is run on any big-endian devices, which is incorrect.

Second, if the player manages to mod in a name that starts with the
noncharacter U+FFFE, then it will appear as FEFF when serialized as
little-endian. This will then fool the decoder into believing that the
rest of the data is actually big-endian and will decode the data as
such, which is of course incorrect.

Explicitly specifying the byte order completely eliminates both of
these problems.